Basic character, adding and removing skill points
Using classes the user is able to add skill points to the various
skills. The program is set up with three accounts, Daniel, Batterink,
and Matt. You are able to add, remove or display the skills of each
character once in each account. The skills include: Strength, Agility,
Intelligence, Resistance, and Dexterity.